requirements.txt
Django
Pillow
django-bootstrap4
django-braces
django-misaka
